# Improve module loading performance