--- a/vuln-reach/src/javascript/lang/imports.rs
+++ b/vuln-reach/src/javascript/lang/imports.rs
@@ -6,7 +6,7 @@ use std::ops::Deref;
 use lazy_static::lazy_static;
 use tree_sitter::{Node, Query, QueryCursor};
 
-use crate::{Error, Tree, JS, Cursor};
+use crate::{Cursor, Error, Tree, JS};
 
 // CommonJS
 //