--- a/vuln-reach/src/javascript/mod.rs
+++ b/vuln-reach/src/javascript/mod.rs
@@ -1,4 +1,3 @@
-mod common;
 pub mod lang;
 pub mod module;
 pub mod package;