This patch addresses some performance issues when creating the access
graphs while new modules are being loaded.

The first performance improvement was a change to the way parent nodes
are accessed in the `AccessGraph::new` function. This repeated node
parent access was taking up the majority of the time in this function
since it would walk the entire tree every time.

To fix this, a new `Cursor` has been added which efficiently finds a
node in a tree and stores all its parents while doing so. While creation
requires walking the tree down to the node, each parent access is just a
simple array lookup.

One minor performance improvement to avoid the time it takes to get to
the target node during `Cursor` creation was adding a `TreeCursorCache`.
This cache stores all the created cursors and allows cloning them
instead of having to walk the tree again. The cache is hit for 50-80% of
all cursor creations, where the percentage seems to increase with bigger
files.

The second and biggest performance improvement was resolving an issue
where modules were loaded multiple times, causing the entire access
graph to be re-created. This could be easily avoided by loading modules
from the existing module cache when they have been encountered before.

Performance improvement samples:
 - `Cursor` sped up `mongoose:7.2.1` from 18s to 3.6s (5x faster)
 - Cache sped up `codemirror:5.5.0` from 33.5s to .8s (42x faster)
 - Combined for `codemirror:5.5.0` from 164s to .8s (206x faster)

Closes #7.